+++ system/src/sheets/PlayerSheetSD.mjs
@@ -654,6 +654,23 @@ export default class PlayerSheetSD extends ActorSheetSD {
 			},
 		]);
 
+		// if equiping shield, remove 2h weapons or switch to 1h
+		if (item.system.isAShield && item.system.equipped) {
+			const updates = [];
+			for (const pi of this.actor.system.getPhysicalItems()) {
+				if (pi.system.isWeapon && pi.system.equipped && pi.system.handedness === "2h") {
+					if (pi.system.isVersatile) {
+						updates.push({ "_id": pi.id, "system.handedness": "1h" });
+					}
+					else {
+						updates.push({ "_id": pi.id, "system.equipped": false });
+					}
+				}
+			}
+			if (updates.length > 0) {
+				await this.actor.updateEmbeddedDocuments("Item", updates);
+			}
+		}
 	}
 
 	async _onToggleHandedness(event) {